+++ suite-common/graph/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/graph-neue",
+    "name": "@suite-common/graph",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
