+++ suite-common/graph/src/constants.ts
