+++ suite-common/graph/src/graphUtils.ts
