+++ suite-common/graph/src/hooks.ts
