+++ suite-common/graph/src/types.ts
