+++ suite-native/module-home/tsconfig.json
@@ -8,9 +8,7 @@
         {
             "path": "../../suite-common/formatters"
         },
-        {
-            "path": "../../suite-common/graph-neue"
-        },
+        { "path": "../../suite-common/graph" },
         {
             "path": "../../suite-common/suite-utils"
         },
